move functions to correct files, add test for empty indexing
